Export llama CI

+++ .ci/scripts/test.sh
@@ -43,14 +43,7 @@ build_cmake_executor_runner() {
   cmake --build ${CMAKE_OUTPUT_DIR} -j4
 }
 
-test_model() {
-  if [[ "${MODEL_NAME}" == "llama2" ]]; then
-    # Test export_llama script: python3 -m examples.models.llama2.export_llama
-    test_export_llama
-  fi
-  # python3 -m examples.portable.scripts.export --model_name="llama2" should works too
-  "${PYTHON_EXECUTABLE}" -m examples.portable.scripts.export --model_name="${MODEL_NAME}"
-
+run_portable_executor_runner() {
   # Run test model
   if [[ "${BUILD_TOOL}" == "buck2" ]]; then
     buck2 run //examples/portable/executor_runner:executor_runner -- --model_path "./${MODEL_NAME}.pte"
@@ -65,23 +58,16 @@ test_model() {
   fi
 }
 
-test_export_llama() {
-  "${PYTHON_EXECUTABLE}" -m examples.models.llama2.export_llama -c examples/models/llama2/params/demo_rand_params.pth -p examples/models/llama2/params/demo_config.json
-
-  # Run test model
-  if [[ "${BUILD_TOOL}" == "buck2" ]]; then
-    buck2 run //examples/portable/executor_runner:executor_runner -- --model_path "./${MODEL_NAME}.pte"
-  elif [[ "${BUILD_TOOL}" == "cmake" ]]; then
-    if [[ ! -f ${CMAKE_OUTPUT_DIR}/executor_runner ]]; then
-      build_cmake_executor_runner
-    fi
-    ./${CMAKE_OUTPUT_DIR}/executor_runner --model_path "./${MODEL_NAME}.pte"
-  else
-    echo "Invalid build tool ${BUILD_TOOL}. Only buck2 and cmake are supported atm"
-    exit 1
+test_model() {
+  if [[ "${MODEL_NAME}" == "llama2" ]]; then
+    # Test export_llama script: python3 -m examples.models.llama2.export_llama
+    "${PYTHON_EXECUTABLE}" -m examples.models.llama2.export_llama -c examples/models/llama2/params/demo_rand_params.pth -p examples/models/llama2/params/demo_config.json
+    run_portable_executor_runner
+    rm "./${MODEL_NAME}.pte"
   fi
-  # clean up
-  rm "./${MODEL_NAME}.pte"
+  # python3 -m examples.portable.scripts.export --model_name="llama2" should works too
+  "${PYTHON_EXECUTABLE}" -m examples.portable.scripts.export --model_name="${MODEL_NAME}"
+  run_portable_executor_runner
 }
 
 build_cmake_xnn_executor_runner() {
